Split monitor.ts to different files
