+++ src/utils/settings.ts
@@ -1,0 +1,11 @@
+import * as vscode from 'vscode';
+
+export function disableDockerEngineAvailabilityPrompt(): void {
+  vscode.workspace
+    .getConfiguration('docker.extension')
+    .update(
+      'dockerEngineAvailabilityPrompt',
+      false,
+      vscode.ConfigurationTarget.Global,
+    );
+}
